Ensure x values of all lines remain in sync
Ensure the x values of all lines remain in sync regardless of whether
they stop and start at any point.

This was done by moving the tracking of the x value out of the
single_line module and into the multi_line module.

The multi_line module now sends an x value along with each new y value
it sends to each single_line it manages.
It also runs a function which prunes all values in each single_line that
fall below the x lookback.